Move debug tile rendering into central utility
