+++ packages/deck.gl-raster/src/index.ts
@@ -1,4 +1,6 @@
 export type { RasterModule } from "./gpu-modules/types.js";
+// Not a public API; exported for use in COGLayer and ZarrLayer
+export { renderDebugTileOutline as _renderDebugTileOutline } from "./layer-utils.js";
 export type { RasterLayerProps, RenderTileResult } from "./raster-layer.js";
 export { RasterLayer } from "./raster-layer.js";
 export type {
